split launcher library functions into their own file
